remove manager file as it was needed
