+++ pkg/gofr/migration/datasource_test.go
@@ -1,12 +1,10 @@
 package migration
 
 import (
-	"testing"
-
 	"github.com/stretchr/testify/assert"
-
 	"gofr.dev/pkg/gofr/container"
 	"gofr.dev/pkg/gofr/testutil"
+	"testing"
 )
 
 func Test_getMigratorDatastoreNotInitialised(t *testing.T) {
@@ -15,7 +13,7 @@ func Test_getMigratorDatastoreNotInitialised(t *testing.T) {
 		mockContainer.SQL = nil
 		mockContainer.Redis = nil
 
-		mg := manager{}
+		mg := Datasource{}
 
 		mg.rollback(mockContainer, transactionData{})
